+++ libexec/got-index-pack/got-index-pack.c
@@ -244,7 +244,8 @@ read_packed_object(struct got_pack *pack, struct got_indexed_object *obj,
 			free(data);
 			break;
 		}
-		if (asprintf(&header, "%s %lld", obj_label, obj->size) == -1) {
+		if (asprintf(&header, "%s %lld", obj_label,
+		    (long long)obj->size) == -1) {
 			err = got_error_from_errno("asprintf");
 			free(data);
 			break;
